arreglando detales de la bd

+++ src/main/java/com/proyecto/Modulo3/repositories/UserRepository.java
@@ -4,7 +4,9 @@ import com.proyecto.Modulo3.entities.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
+import java.util.List;
+
 @Repository
-public interface UserRepository extends JpaRepository<User, Long> {
+public interface UserRepository extends JpaRepository<User, Integer> {
 
 }
