+++ src/main/java/com/proyecto/Modulo3/services/UserDetailServices.java
@@ -1,7 +1,6 @@
 package com.proyecto.Modulo3.services;
 
 import com.proyecto.Modulo3.dto.UserDetailDTO;
-import org.springframework.stereotype.Service;
 
 import java.util.List;
 
@@ -9,11 +8,11 @@ public interface UserDetailServices {
 
     List<UserDetailDTO> listUserDetails();
 
-    UserDetailDTO getUserDetail(Long id);
+    UserDetailDTO getUserDetail(Integer id);
 
     UserDetailDTO createUserDetail(UserDetailDTO userDetail);
 
-    UserDetailDTO updateUserDetail(Long id, UserDetailDTO userDetail);
+    UserDetailDTO updateUserDetail(Integer id, UserDetailDTO userDetail);
 
-    Long deleteUserDetail(Long id);
+    Integer deleteUserDetail(Integer id);
 }
